Moved Sidebar into component
