+++ src/components/sidebar/index.ts
@@ -1,5 +1,9 @@
+import Sidebar from './sidebar.svelte'
+
 export type NavigationElement = {
 	name: string
 	icon: ConstructorOfATypedSvelteComponent
 	route: string
 }
+
+export { Sidebar }
